refactor(ActionMenu): Extract code in some components
to better update them later. No code change
