+++ react/ActionMenu/ActionMenuRadio.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+import Radio from '../Radio'
+
+import styles from './styles.styl'
+
+export const ActionMenuRadio = props => {
+  return <Radio {...props} className={styles['c-actionmenu-radio']} />
+}
